perf test cleanup
